merges translator class with IO file, confirms functionality of previous methods.
