+++ message.txt
@@ -1,3 +1,1 @@
-1234
-Good morning, Denver!
-BEEEEEP
+t
